fix(llm): preserve additionalProperties on OpenAI-compat tool schemas
- Extract shared strict schema normalizer to services/tool_schema_normalize.py
- Use OpenAI-specific normalizer for agent and chat OpenAI-compat paths
- Keep Bedrock Converse stripping additionalProperties via bedrock key policy
- Add regression tests for provider-specific normalization behavior
